+++ XCDYouTubeKit Tests/XCDYouTubeProtectedVideosTestCase.m
@@ -33,6 +33,28 @@
 	[self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
+- (void) testAgeRestrictedUnratedVideo
+{
+	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];
+	[[XCDYouTubeClient defaultClient] getVideoWithIdentifier:@"KGZzuVNwJHY" completionHandler:^(XCDYouTubeVideo *video, NSError *error)
+	{
+		XCTAssertNil(error);
+		XCTAssertNotNil(video.title);
+		XCTAssertNotNil(video.expirationDate);
+		XCTAssertNotNil(video.smallThumbnailURL);
+		XCTAssertNotNil(video.mediumThumbnailURL);
+		XCTAssertNotNil(video.largeThumbnailURL);
+		XCTAssertTrue(video.streamURLs.count > 0);
+		XCTAssertTrue(video.duration > 0);
+		[video.streamURLs enumerateKeysAndObjectsUsingBlock:^(NSNumber *key, NSURL *streamURL, BOOL *stop)
+		{
+			XCTAssertTrue([streamURL.query rangeOfString:@"signature="].location != NSNotFound);
+		}];
+		[expectation fulfill];
+	}];
+	[self waitForExpectationsWithTimeout:5 handler:nil];
+}
+
 - (void) testProtectedVEVOVideo1
 {
 	__weak XCTestExpectation *expectation = [self expectationWithDescription:@""];