+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### Version 2.1.2
 
 * Fixed playback of some protected videos. (#122)
+* Fixed playback of some age restricted videos. (#137)
 
 #### Version 2.1.1